requirements.txt
paddleocr>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pypdfium2>=4.0.0

